issue: 3514044 Moving cq_mgr_regrq tx methods to cq_mgr
