+++ src/core/dev/cq_mgr.cpp
@@ -48,6 +48,7 @@
 #include "buffer_pool.h"
 #include "qp_mgr.h"
 #include "ring_simple.h"
+#include "qp_mgr_eth_mlx5.h"
 
 #define MODULE_NAME "cqm"
 
@@ -72,7 +73,8 @@ uint64_t cq_mgr::m_n_global_sn = 0;
 
 cq_mgr::cq_mgr(ring_simple *p_ring, ib_ctx_handler *p_ib_ctx_handler, int cq_size,
                struct ibv_comp_channel *p_comp_event_channel, bool is_rx, bool config)
-    : m_p_ibv_cq(NULL)
+    : m_qp(NULL)
+    , m_p_ibv_cq(NULL)
     , m_b_is_rx(is_rx)
     , m_cq_id(0)
     , m_n_cq_poll_sn(0)
@@ -114,6 +116,8 @@ cq_mgr::cq_mgr(ring_simple *p_ring, ib_ctx_handler *p_ib_ctx_handler, int cq_siz
     if (config) {
         configure(cq_size);
     }
+
+    memset(&m_mlx5_cq, 0, sizeof(m_mlx5_cq));
 }
 
 void cq_mgr::configure(int cq_size)
@@ -713,31 +717,6 @@ int cq_mgr::poll_and_process_element_rx(uint64_t *p_cq_poll_sn, void *pv_fd_read
     return ret_rx_processed;
 }
 
-int cq_mgr::poll_and_process_element_tx(uint64_t *p_cq_poll_sn)
-{
-    // Assume locked!!!
-    cq_logfuncall("");
-
-    /* coverity[stack_use_local_overflow] */
-    xlio_ibv_wc wce[MCE_MAX_CQ_POLL_BATCH];
-    int ret = poll(wce, m_n_sysvar_cq_poll_batch_max, p_cq_poll_sn);
-    if (ret > 0) {
-        m_n_wce_counter += ret;
-        if (ret < (int)m_n_sysvar_cq_poll_batch_max) {
-            m_b_was_drained = true;
-        }
-
-        for (int i = 0; i < ret; i++) {
-            mem_buf_desc_t *buff = cqe_log_and_get_buf_tx((&wce[i]));
-            if (buff) {
-                process_tx_buffer_list(buff);
-            }
-        }
-    }
-
-    return ret;
-}
-
 bool cq_mgr::reclaim_recv_buffers(mem_buf_desc_t *rx_reuse_lst)
 {
     if (m_rx_buffs_rdy_for_free_head) {
@@ -1023,4 +1002,96 @@ cq_mgr *get_cq_mgr_from_cq_event(struct ibv_comp_channel *p_cq_channel)
     ENDIF_VERBS_FAILURE;
 
     return p_cq_mgr;
+}
+
+int cq_mgr::poll_and_process_element_tx(uint64_t *p_cq_poll_sn)
+{
+    cq_logfuncall("");
+
+    static auto is_error_opcode = [&](uint8_t opcode) {
+        return opcode == MLX5_CQE_REQ_ERR || opcode == MLX5_CQE_RESP_ERR;
+    };
+
+    int ret = 0;
+    uint32_t num_polled_cqes = 0;
+    xlio_mlx5_cqe *cqe = get_cqe_tx(num_polled_cqes);
+
+    if (likely(cqe)) {
+        unsigned index = ntohs(cqe->wqe_counter) & (m_qp->m_tx_num_wr - 1);
+
+        // All error opcodes have the most significant bit set.
+        if (unlikely(cqe->op_own & 0x80) && is_error_opcode(cqe->op_own >> 4)) {
+            m_p_cq_stat->n_rx_cqe_error++;
+            log_cqe_error(cqe);
+        }
+
+        handle_sq_wqe_prop(index);
+        ret = 1;
+    }
+    update_global_sn(*p_cq_poll_sn, num_polled_cqes);
+
+    return ret;
+}
+
+void cq_mgr::log_cqe_error(struct xlio_mlx5_cqe *cqe)
+{
+    struct mlx5_err_cqe *ecqe = (struct mlx5_err_cqe *)cqe;
+
+    /* TODO We can also ask qp_mgr to log WQE fields from SQ. But at first, we need to remove
+     * prefetch and memset of the next WQE there. Credit system will guarantee that we don't
+     * reuse the WQE at this point.
+     */
+
+    if (MLX5_CQE_SYNDROME_WR_FLUSH_ERR != ecqe->syndrome) {
+        cq_logwarn("cqe: syndrome=0x%x vendor=0x%x hw=0x%x (type=0x%x) wqe_opcode_qpn=0x%x "
+                   "wqe_counter=0x%x",
+                   ecqe->syndrome, ecqe->vendor_err_synd, *((uint8_t *)&ecqe->rsvd1 + 16),
+                   *((uint8_t *)&ecqe->rsvd1 + 17), ntohl(ecqe->s_wqe_opcode_qpn),
+                   ntohs(ecqe->wqe_counter));
+    }
+}
+
+void cq_mgr::handle_sq_wqe_prop(unsigned index)
+{
+    sq_wqe_prop *p = &m_qp->m_sq_wqe_idx_to_prop[index];
+    sq_wqe_prop *prev;
+    unsigned credits = 0;
+
+    /*
+     * TX completions can be signalled for a set of WQEs as an optimization.
+     * Therefore, for every TX completion we may need to handle multiple
+     * WQEs. Since every WQE can have various size and the WQE index is
+     * wrapped around, we build a linked list to simplify things. Each
+     * element of the linked list represents properties of a previously
+     * posted WQE.
+     *
+     * We keep index of the last completed WQE and stop processing the list
+     * when we reach the index. This condition is checked in
+     * is_sq_wqe_prop_valid().
+     */
+
+    do {
+        if (p->buf) {
+            m_p_ring->mem_buf_desc_return_single_locked(p->buf);
+        }
+        if (p->ti) {
+            xlio_ti *ti = p->ti;
+            if (ti->m_callback) {
+                ti->m_callback(ti->m_callback_arg);
+            }
+
+            ti->put();
+            if (unlikely(ti->m_released && ti->m_ref == 0)) {
+                m_qp->ti_released(ti);
+            }
+        }
+        credits += p->credits;
+
+        prev = p;
+        p = p->next;
+    } while (p != NULL && m_qp->is_sq_wqe_prop_valid(p, prev));
+
+    m_p_ring->return_tx_pool_to_global_pool();
+    m_qp->credits_return(credits);
+    m_qp->m_sq_wqe_prop_last_signalled = index;
 }
